% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_double_prime_new}
\alias{f_double_prime_new}
\title{Title}
\usage{
f_double_prime_new(tau, tau_hat, varsigma, xi, kappa, R)
}
\arguments{
\item{tau}{}

\item{tau_hat}{}

\item{varsigma}{}

\item{xi}{}

\item{kappa}{}

\item{R}{}
}
\value{

}
\description{
Title
}
