% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{g.eta.double.prime}
\alias{g.eta.double.prime}
\title{Title}
\usage{
g.eta.double.prime(tau, tau.hat, varsigma, xi, eta.hat, eps)
}
\arguments{
\item{tau}{}

\item{tau.hat}{}

\item{varsigma}{}

\item{xi}{}

\item{eta.hat}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
