% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g_eta_double_prime_new}
\alias{g_eta_double_prime_new}
\title{Title}
\usage{
g_eta_double_prime_new(tau, tau_hat, varsigma, xi, eta_hat, eps)
}
\arguments{
\item{tau}{}

\item{tau_hat}{}

\item{varsigma}{}

\item{xi}{}

\item{eta_hat}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
