% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatGEVBMA.wrapper.R
\name{get.sigma.22.inv}
\alias{get.sigma.22.inv}
\title{Title}
\usage{
get.sigma.22.inv(R, burn = NULL, odens = 1000)
}
\arguments{
\item{R}{}

\item{burn}{}

\item{odens}{}
}
\value{

}
\description{
Title
}
