% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatGEVBMA.wrapper.R
\name{gev.impute.params}
\alias{gev.impute.params}
\title{Title}
\usage{
gev.impute.params(
  R,
  X.drop,
  S.drop,
  sigma.22.inv,
  sigma.22.inv.tau,
  burn = NULL,
  xi.constrain = c(-Inf, Inf),
  odens = 1000
)
}
\arguments{
\item{R}{}

\item{X.drop}{}

\item{S.drop}{}

\item{sigma.22.inv}{}

\item{sigma.22.inv.tau}{}

\item{burn}{}

\item{xi.constrain}{}

\item{odens}{}
}
\value{

}
\description{
Title
}
