\name{gev.results.init}
\alias{gev.results.init}

\title{Initialize a results object for spatial.bma.gev

}
\description{This internal function initializes the object that stores
  the results of an MCMC run in spatial.bma.gev

}
\usage{
gev.results.init(n.s, p, n.reps)
}

\arguments{
  \item{n.s}{The number of sites

}
  \item{p}{The number of covariates in X

}
  \item{n.reps}{The total number of repetitions

}
}

\value{
 A object that will store the results.
}

\author{
  Alex Lenkoski <alex@nr.no>
}
