\name{gev.update.hyper}
\alias{gev.update.hyper}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Updates the Gaussian Process hyperparameters in the Spatial GEV model

}
\description{This function is used internally to update the
  Gaussian Process hyperparameters of a Spatial GEV model.
}
\usage{
gev.update.hyper(G)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{G}{The current state of a Spatial GEV model

}
}
\value{
An updated state of a Spatial GEV model.
}
\author{
Alex Lenkoski <alex@nr.no>
}
