\name{gev.update}
\alias{gev.update}

\title{Updates all the parameters in a spatial GEV model

}
\description{This internal function updates all the components of a
  spatial GEV model

}
\usage{
gev.update(G)
}
\arguments{
  \item{G}{The current state of a spatial GEV model

}
}

\value{
  This returns an updated state
}

\author{
  Alex Lenkoski <alex@nr.no>
}
