% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gevUpdateM}
\alias{gevUpdateM}
\title{Title}
\usage{
gevUpdateM(Y, X, M, alpha, lambda, D, beta0, Omega0)
}
\arguments{
\item{Y}{}

\item{X}{}

\item{M}{}

\item{alpha}{}

\item{lambda}{}

\item{D}{}

\item{beta0}{}

\item{Omega0}{}
}
\value{

}
\description{
Title
}
