% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatGEVBMA.wrapper.R
\name{imputation.func}
\alias{imputation.func}
\title{Title}
\usage{
imputation.func(
  i,
  cov.map,
  S.map,
  R,
  sigma.22.inv,
  sigma.22.inv.tau,
  return.period,
  all.post.quantiles,
  N,
  xi.constrain = c(-Inf, Inf)
)
}
\arguments{
\item{i}{}

\item{cov.map}{}

\item{S.map}{}

\item{R}{}

\item{sigma.22.inv}{}

\item{sigma.22.inv.tau}{}

\item{return.period}{}

\item{all.post.quantiles}{}

\item{N}{}

\item{xi.constrain}{}
}
\value{

}
\description{
Title
}
