% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{j_prime_new}
\alias{j_prime_new}
\title{Title}
\usage{
j_prime_new(tau, tau_hat, varsigma, kappa, xi_hat, eps)
}
\arguments{
\item{tau}{}

\item{tau_hat}{}

\item{varsigma}{}

\item{kappa}{}

\item{xi_hat}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
