% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldot}
\alias{ldot}
\title{Title}
\usage{
ldot(tau, alpha, lambda, D, a, b)
}
\arguments{
\item{tau}{}

\item{alpha}{}

\item{lambda}{}

\item{D}{}

\item{a}{}

\item{b}{}
}
\value{

}
\description{
Title
}
