% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_spatgev.R
\name{make_temporal_spatgev_data}
\alias{make_temporal_spatgev_data}
\title{Make input data to spatial.gev.bma() when having year-specific covariates}
\usage{
make_temporal_spatgev_data(amax_data, to_standardize = TRUE)
}
\arguments{
\item{amax_data}{A data table with columns year, stid (location/station identification number) and y (the annual max observation of e.g. hourly precip), in addition to other columns that include covariate data.}

\item{to_standardize}{TRUE if we should standardize the covariates.}
}
\value{
List of Y, X and S that can be used as input in spatial.gev.bma().
}
\description{
Make input data to spatial.gev.bma() when having year-specific covariates
}
\examples{
setwd("/nr/samba/user/roksvag/GitRepo/SpatGEVBMA/scripts/dev_I4C/")
amax_data=fread(file="Data/AM60.csv")
spatgev_data=make_temporal_spatgev_data(amax_data,TRUE)
spatial.gev.bma(Y.list=spatgev_data$Y,X.all=spatgev_data$X,S=spatgev_data$S,n.reps=1000)
}
