\name{norway}
\docType{data}
\alias{norway}
\title{Extreme Precipitation Data at 69 Sites in Norway}
\description{
  This dataset provides observations of extreme precipitation in Norway
  at 69 sites, along with covariate information.
}
\usage{data(norway)}
\format{A list containing three objects.  Y.list is itself a list, with
  the observations at each site.  X is a matrix with the coviates and S
  is a matrix with their relative position}
\keyword{datasets}