\name{SpatGEVBMA-package}
\alias{SpatGEVBMA-package}
\alias{SpatGEVBMA}
\docType{package}
\title{\packageTitle{SpatGEVBMA}}
\description{\packageDescription{SpatGEVBMA}}
\section{Package Content}{\packageIndices{SpatGEVBMA}}
\author{\packageAuthor{SpatGEVBMA}}
\section{Maintainer}{\packageMaintainer{SpatGEVBMA}}
\keyword{package}
