\name{spatial.gev.bma-package}
\alias{spatial.gev.bma-package}
\docType{package}
\title{
Fit a Hierarchical Spatial Generalized Extreme Value model that allows
for Bayesian Model Averaging
}
\description{
This package allows one to fit a hierarhical spatial model for
extremes.  It enables linear specifications in all three GEV parameters
and incorporates Bayesian Model Averaging to account for model
uncertainty.  Further, it uses recent results in automatic proposal
formation of Metropolis-Hastings samplers to avoid requiring the user to
have to specify a large number of tuning parameters.
}
\details{
\tabular{ll}{
Package: \tab spatial.gev.bma\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-11\cr
License: \tab GPL\cr
}
}
\author{
Alex Lenkoski wrote the software for a specific project.  He's more than
happy to hear from you if you'd like to see additional features added.

Maintainer: Alex Lenkoski <alex@nr.no>

}
\references{
A. Dyrrdal, A. Lenkoski, T. Thorarinsdottir, F. Stordal (2014).
Bayesian hierarchical modeling of extreme hourly precipitation in
Norway. arxiv:1309.6111
}
\keyword{ package }

